tuesday night

+++ server.py
@@ -32,6 +32,9 @@ def list_page():
 
 @app.route("/search")
 def search():
+    email = None
+    if 'email' in session:
+        email = escape(session['email'])
     search_phrase = request.args['q'].lower()
     if search_phrase:
         answers_with_sp = data_handler.get_answers_with_sp(search_phrase)
@@ -42,7 +45,7 @@ def search():
         answers_with_sp = list(map(lambda record: modifier(record), answers_with_sp))
         question_list = data_handler.search(search_phrase)
         question_list = list(map(lambda record: modifier(record), question_list))
-        return render_template("list.html", question_list=question_list, answers_with_sp=answers_with_sp)
+        return render_template("list.html", question_list=question_list, answers_with_sp=answers_with_sp, email=email)
     return redirect('/')
 
 
@@ -71,109 +74,148 @@ def question_page(id=None):
 
 @app.route("/question/<question_id>/edit", methods=['GET', 'POST'])
 def edit(question_id):
-    if request.method == 'POST':
-        data_handler.edit_question(question_id, request.form['title'], request.form['message'])
-        return redirect('/question/' + question_id)
-    question = data_handler.get_one_by_id("question", question_id)
-    return render_template('edit_question.html', question=question)
+    email = None
+    if 'email' in session:
+        email_by_question_id = data_handler.get_one_by_id('question', question_id)['email']
+        email = escape(session['email'])
+        if email == email_by_question_id:
+            if request.method == 'POST':
+                data_handler.edit_question(question_id, request.form['title'], request.form['message'])
+                return redirect('/question/' + question_id)
+            question = data_handler.get_one_by_id("question", question_id)
+            return render_template('edit_question.html', question=question, email=email)
+    return redirect('/')
 
 
 @app.route("/answer/<answer_id>/edit", methods=['GET', 'POST'])
 def edit_answer(answer_id):
-    answer_record = data_handler.get_one_by_id("answer", answer_id)
-    if request.method == 'POST':
-        data_handler.edit_answer(answer_id, request.form['message'])
-        return redirect('/question/' + str(answer_record["question_id"]))
-    answer = data_handler.get_one_by_id("answer", answer_id)
-    return render_template('edit_answer.html', answer=answer)
+    email = None
+    if 'email' in session:
+        email_by_answer_id = data_handler.get_one_by_id('answer', answer_id)['email']
+        email = escape(session['email'])
+        if email == email_by_answer_id:
+            answer_record = data_handler.get_one_by_id("answer", answer_id)
+            if request.method == 'POST':
+                data_handler.edit_answer(answer_id, request.form['message'])
+                return redirect('/question/' + str(answer_record["question_id"]))
+            answer = data_handler.get_one_by_id("answer", answer_id)
+            return render_template('edit_answer.html', answer=answer, email=email)
+    return redirect('/')
 
 
 @app.route("/comment/<comment_id>/edit", methods=['GET', 'POST'])
 def edit_comment(comment_id):
-    comment_record = data_handler.get_one_by_id("comment", comment_id)
-    question_id = comment_record['question_id']
-    if comment_record['answer_id']:
-        question_id = data_handler.get_one_by_id("answer", comment_record['answer_id'])['question_id']
-    if request.method == 'POST':
-        data_handler.edit_comment(comment_id, request.form['message'])
-        data_handler.increase_comment_edit(comment_id)
-        return redirect('/question/' + str(question_id))
-    return render_template('edit_comment.html', comment=comment_record)
-
+    email = None
+    if 'email' in session:
+        email_by_comment_id = data_handler.get_one_by_id('comment', comment_id)['email']
+        email = escape(session['email'])
+        if email == email_by_comment_id:
+            comment_record = data_handler.get_one_by_id("comment", comment_id)
+            question_id = comment_record['question_id']
+            if comment_record['answer_id']:
+                question_id = data_handler.get_one_by_id("answer", comment_record['answer_id'])['question_id']
+            if request.method == 'POST':
+                data_handler.edit_comment(comment_id, request.form['message'])
+                data_handler.increase_comment_edit(comment_id)
+                return redirect('/question/' + str(question_id))
+            return render_template('edit_comment.html', comment=comment_record, email=email)
+    return redirect('/')
 
 @app.route("/add-question", methods=["GET", "POST"])
 def add_question():
-    if 'email' in session: 
-        user_id = data_handler.get_record_by_email(session['email'])['id']
-    if request.method == "POST":
-        question_id = data_handler.new_question(request.form, user_id)
-        f = request.files['picture_upload']
-        pic_name = ""
-        if f.filename:
-            pic_name = "images/uploaded_images/" + secure_filename(str(question_id) + "." + f.content_type.split("/")[1])
-            f.save("./" + pic_name)
-            data_handler.insert_picture("question", question_id, pic_name)
-        return redirect("/question/" + str(question_id))
-    return render_template("add_question.html")
+    email = None
+    if 'email' in session:
+        user_id = data_handler.get_record_by_email(session['email'])['userid']
+        email = escape(session['email'])
+        if request.method == "POST":
+            question_id = data_handler.new_question(request.form, user_id)
+            f = request.files['picture_upload']
+            pic_name = ""
+            if f.filename:
+                pic_name = "images/uploaded_images/" + secure_filename(str(question_id) + "." + f.content_type.split("/")[1])
+                f.save("./" + pic_name)
+                data_handler.insert_picture("question", question_id, pic_name)
+            return redirect("/question/" + str(question_id))
+        return render_template("add_question.html", email=email)
+    return redirect('/')
 
 
 @app.route("/question/<question_id>/new-answer", methods=["GET", "POST"])
 def add_new_answer(question_id=None):
     if 'email' in session:
-        user_id = data_handler.get_record_by_email(session['email'])['id']
-    answer_id = data_handler.new_answer(question_id, request.form, user_id)
-    f = request.files['picture_upload']
-    pic_name = ""
-    if f.filename:
-        pic_name = "images/uploaded_images/answer/" + secure_filename(str(answer_id) + "." + f.content_type.split("/")[1])
-        f.save("./" + pic_name)
-        data_handler.insert_picture("answer", answer_id, pic_name)
-    return redirect("/question/" + str(question_id))
+        user_id = data_handler.get_record_by_email(session['email'])['userid']
+        answer_id = data_handler.new_answer(question_id, request.form, user_id)
+        f = request.files['picture_upload']
+        pic_name = ""
+        if f.filename:
+            pic_name = "images/uploaded_images/answer/" + secure_filename(str(answer_id) + "." + f.content_type.split("/")[1])
+            f.save("./" + pic_name)
+            data_handler.insert_picture("answer", answer_id, pic_name)
+        return redirect("/question/" + str(question_id))
+    return redirect('/')
 
 
 @app.route("/answer/<answer_id>/delete")
 def delete_answer(answer_id=None):
-    answer_record = data_handler.get_one_by_id("answer", answer_id)
-    util.delete_pictures([] + [answer_record])
-    data_handler.delete_record("answer", "id", answer_id)
-    return redirect("/question/" + str(answer_record['question_id']))
+    if 'email' in session:
+        email_by_answer_id = data_handler.get_one_by_id('answer', answer_id)['email']
+        email = escape(session['email'])
+        if email == email_by_answer_id:
+            answer_record = data_handler.get_one_by_id("answer", answer_id)
+            util.delete_pictures([] + [answer_record])
+            data_handler.delete_record("answer", "id", answer_id)
+            return redirect("/question/" + str(answer_record['question_id']))
+    return redirect('/')
 
 
 @app.route("/question/<question_id>/delete")
 def delete_question(question_id):
-    question_record = data_handler.get_one_by_id("question", question_id)
-    answer_records = data_handler.get_many_by_id("answer", "question_id", question_id)
-    util.delete_pictures(answer_records + [question_record])
-    data_handler.delete_record("question_tag", "question_id", question_id)
-    data_handler.delete_record("comment", "question_id", question_id)
-    for answer in answer_records:
-        data_handler.delete_record("comment", "answer_id", answer["id"])
-    data_handler.delete_record("answer", "question_id", question_id)
-    data_handler.delete_record("question", "id", question_id)
-    return redirect("/")
+    if 'email' in session:
+        email_by_question_id = data_handler.get_one_by_id('question', question_id)['email']
+        email = escape(session['email'])
+        if email == email_by_question_id:
+            question_record = data_handler.get_one_by_id("question", question_id)
+            answer_records = data_handler.get_many_by_id("answer", "question_id", question_id)
+            util.delete_pictures(answer_records + [question_record])
+            data_handler.delete_record("question_tag", "question_id", question_id)
+            data_handler.delete_record("comment", "question_id", question_id)
+            for answer in answer_records:
+                data_handler.delete_record("comment", "answer_id", answer["id"])
+            data_handler.delete_record("answer", "question_id", question_id)
+            data_handler.delete_record("question", "id", question_id)
+    return redirect("/list")
+            
 
 
 @app.route("/question/<id>/new-comment", methods=["POST", "GET"])
 def new_comment_question(id):
+    email = None
     if 'email' in session:
-        user_id = data_handler.get_record_by_email(session['email'])['id']
-    if request.method == "GET":
-        return render_template("new-comment.html", id=id, record_type="question")
-    new_comment = request.form['message']
-    data_handler.new_comment(new_comment, user_id, question_id=id)
-    return redirect('/question/' + str(id))
+        user_id = data_handler.get_record_by_email(session['email'])['userid']
+        email = escape(session['email'])
+        if request.method == "GET":
+            return render_template("new-comment.html", id=id, record_type="question", email=email)
+        else:
+            new_comment = request.form['message']
+            data_handler.new_comment(new_comment, user_id, question_id=id)
+            return redirect('/question/' + str(id))
+    return redirect('/')
 
 
 @app.route("/answer/<id>/new-comment", methods=["POST", "GET"])
 def new_comment_answer(id):
+    email = None
     if 'email' in session:
-        user_id = data_handler.get_record_by_email(session['email'])['id']
-    if request.method == "GET":
-        return render_template("new-comment.html", id=id, record_type="answer")
-    new_comment = request.form['message']
-    data_handler.new_comment(new_comment, user_id, answer_id=id)
-    answer_record = data_handler.get_one_by_id('answer', id)
-    return redirect('/question/' + str(answer_record['question_id']))
+        user_id = data_handler.get_record_by_email(session['email'])['userid']
+        email = escape(session['email'])
+        if request.method == "GET":
+            return render_template("new-comment.html", id=id, record_type="answer", email=email)
+        else:
+            new_comment = request.form['message']
+            data_handler.new_comment(new_comment, user_id, answer_id=id)
+            answer_record = data_handler.get_one_by_id('answer', id)
+            return redirect('/question/' + str(answer_record['question_id']))
+    return redirect('/')
 
 
 @app.route("/<record_type>/<id>/<vote_type>")
@@ -197,29 +239,44 @@ def open_up_picture(picture_type, id):
 
 @app.route("/question/<question_id>/new-tag", methods=["GET", "POST"])
 def new_tag(question_id):
-    existing_tags = data_handler.get_tags()
-    if request.method == "GET":
-        return render_template("add_tag.html", id=question_id, existing_tags=existing_tags)
-    new_tag = request.form['tag_name'] if request.form['tag_name'] else request.form['existing_tag']
-    data_handler.add_new_tag(question_id, new_tag)
-    return redirect("/question/" + str(question_id))
+    email = None
+    if 'email' in session:
+        email_by_question_id = data_handler.get_one_by_id('question', question_id)['email']
+        email = escape(session['email'])
+        if email == email_by_question_id:
+            existing_tags = data_handler.get_tags()
+            if request.method == "GET":
+                return render_template("add_tag.html", id=question_id, existing_tags=existing_tags, email=email)
+            new_tag = request.form['tag_name'] if request.form['tag_name'] else request.form['existing_tag']
+            data_handler.add_new_tag(question_id, new_tag)
+            return redirect("/question/" + str(question_id))
+    return redirect('/')
 
 
 @app.route("/question/<question_id>/tag/<tag_id>/delete")
 def delete_tag(question_id=None, tag_id=None):
-    data_handler.delete_tag(question_id, tag_id)
-    return redirect("/question/" + str(question_id))
+    if 'email' in session:
+        email_by_question_id = data_handler.get_one_by_id('question', question_id)['email']
+        email = escape(session['email'])
+        if email == email_by_question_id:
+            data_handler.delete_tag(question_id, tag_id)
+            return redirect("/question/" + str(question_id))
+    return redirect('/')
 
 
 @app.route("/comments/<comment_id>/delete")
 def delete_comment(comment_id):
-    print(comment_id)
-    comment_record = data_handler.get_one_by_id("comment", comment_id)
-    question_id = comment_record['question_id']
-    if comment_record['answer_id']:
-        question_id = data_handler.get_one_by_id("answer", comment_record['answer_id'])['question_id']
-    data_handler.delete_comment(comment_id)
-    return redirect('/question/' + str(question_id))
+    if 'email' in session:
+        email_by_question_id = data_handler.get_one_by_id('question', question_id)['email']
+        email = escape(session['email'])
+        if email == email_by_question_id:
+            comment_record = data_handler.get_one_by_id("comment", comment_id)
+            question_id = comment_record['question_id']
+            if comment_record['answer_id']:
+                question_id = data_handler.get_one_by_id("answer", comment_record['answer_id'])['question_id']
+            data_handler.delete_comment(comment_id)
+            return redirect('/question/' + str(question_id))
+    return redirect('/')
 
 
 @app.route("/registration", methods=['GET', 'POST'])
